+++ src/Storage.ts
@@ -3,6 +3,7 @@
  */
 import localForage from "localforage";
 import { instanceToPlain, plainToInstance } from 'class-transformer';
+import { isFinite } from "lodash";
 import { mastodon } from "masto";
 
 import Account from "./api/objects/account";
@@ -19,8 +20,8 @@ import { checkUniqueIDs, zipPromises } from "./helpers/collection_helpers";
 import { config } from "./config";
 import { DEFAULT_WEIGHTS } from "./scorer/weight_presets";
 import { isDebugMode, isDeepDebug } from "./helpers/environment_helpers";
-import { isNumber, sizeOf } from "./helpers/math_helper";
 import { Logger } from './helpers/logger';
+import { sizeOf } from "./helpers/math_helper";
 import {
     type FeedFilterSettings,
     type FeedFilterSettingsSerialized,
@@ -186,7 +187,7 @@ export default class Storage {
         Object.entries(DEFAULT_WEIGHTS).forEach(([key, defaultValue]) => {
             const value = weights[key as WeightName]
 
-            if (!isNumber(value)) {
+            if (!isFinite(value)) {
                 logger.warn(`Missing value for "${key}" in saved weights, setting to default: ${defaultValue}`);
                 weights[key as WeightName] = DEFAULT_WEIGHTS[key as WeightName];
                 shouldSave = true;