+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # NEXT RELEASE
-* Get rid of skipSort param; introduce TootSource type
+* Get rid of `skipSort` param; introduce `TootSource` type
 
 ### v0.69.13
 * Set `FEDIVERSE_POPULAR_SERVERS` staleness to 5 days